Extract demo-mode support specific code
