+++ lib/main.js
@@ -4,7 +4,7 @@ const Base = require("./base")
 const globalState = require("./global-state")
 const settings = require("./settings")
 const VimState = require("./vim-state")
-
+let DemoModeSupport
 let paneUtils
 
 function getPaneUtils() {
@@ -139,6 +139,7 @@ module.exports = {
 
   deactivate() {
     this.demaximizePane()
+    if (this.demoModeSupport) this.demoModeSupport.destroy()
 
     this.subscriptions.dispose()
     VimState.forEach(vimState => vimState.destroy())
@@ -218,7 +219,6 @@ module.exports = {
   registerVimStateCommands() {
     // all commands here is executed with context where 'this' bound to 'vimState'
     // prettier-ignore
-
     const commands = {
       "activate-normal-mode"() { this.activate("normal") },
       "activate-linewise-visual-mode"() { this.activate("visual", "linewise") },
@@ -282,39 +282,9 @@ module.exports = {
     this.subscribe(new Disposable(() => statusBarManager.detach()))
   },
 
-  consumeDemoMode({onWillAddItem, onDidStart, onDidStop, onDidRemoveHover}) {
-    this.subscribe(
-      onDidStart(() => globalState.set("demoModeIsActive", true)),
-      onDidStop(() => globalState.set("demoModeIsActive", false)),
-      onDidRemoveHover(this.destroyAllDemoModeFlasheMarkers.bind(this)),
-      onWillAddItem(({item, event}) => {
-        if (event.binding.command.startsWith("vim-mode-plus:")) {
-          const commandElement = item.getElementsByClassName("command")[0]
-          commandElement.textContent = commandElement.textContent.replace(/^vim-mode-plus:/, "")
-        }
-        const element = document.createElement("span")
-        element.classList.add("kind", "pull-right")
-        element.textContent = this.getKindForCommand(event.binding.command)
-        item.appendChild(element)
-      })
-    )
-  },
-
-  destroyAllDemoModeFlasheMarkers() {
-    VimState.forEach(vimState => vimState.flashManager.destroyDemoModeMarkers())
-  },
-
-  getKindForCommand(command) {
-    if (command.startsWith("vim-mode-plus")) {
-      command = command.replace(/^vim-mode-plus:/, "")
-      if (command.startsWith("operator-modifier")) {
-        return "op-modifier"
-      } else {
-        return Base.getKindForCommandName(command) || "vmp-other"
-      }
-    } else {
-      return "non-vmp"
-    }
+  consumeDemoMode(demoModeService) {
+    if (!DemoModeSupport) DemoModeSupport = require("./demo-mode-support")
+    this.demoModeSupport = new DemoModeSupport(demoModeService)
   },
 
   createVimState(editor) {
